Change to one-window GUI, all widgets now embedded; move camera addons to one file for now; buffer to be recorded not in Recording widget anymore > should be moved to Startup setup;
