+++ Startup.py
@@ -154,6 +154,155 @@ class CameraWidget(QtWidgets.QGroupBox):
                 self._listw_cameras.addItem(name)
 
 
+################################
+# Recordings widget (SALVAGED FROM GUI INTEGRATED)
+
+class Recording(QtWidgets.QGroupBox):
+
+    def __init__(self, _main):
+        QtWidgets.QGroupBox.__init__(self, 'Recordings')
+        self._main : GUI.Main = _main
+
+        vSpacer = QtWidgets.QSpacerItem(1,1, QtWidgets.QSizePolicy.Minimum, QtWidgets.QSizePolicy.Expanding)
+
+        self.setLayout(QtWidgets.QHBoxLayout())
+        ### Create inner widget
+        self.wdgt = QtWidgets.QWidget(self)
+        self.wdgt.setObjectName('InGroupBox')
+        self.wdgt.setLayout(QtWidgets.QGridLayout())
+        self.layout().addWidget(self.wdgt)
+
+        ### Basic properties
+        self.setCheckable(True)
+
+        ### Current folder
+        self._le_folder = QtWidgets.QLineEdit()
+        self._le_folder.setEnabled(False)
+        self.wdgt.layout().addWidget(QtWidgets.QLabel('Folder'), 0, 0)
+        self.wdgt.layout().addWidget(self._le_folder, 0, 1, 1, 2)
+
+        ### GroupBox
+        self.clicked.connect(self.toggleEnable)
+
+        ### Buttons
+        ## Start
+        self._btn_start = QtWidgets.QPushButton('Start')
+        self._btn_start.clicked.connect(lambda: IPC.rpc(Def.Process.Controller, Controller.startRecording))
+        self.wdgt.layout().addWidget(self._btn_start, 1, 0)
+        ## Pause
+        self._btn_pause = QtWidgets.QPushButton('Pause')
+        self._btn_pause.clicked.connect(lambda: IPC.rpc(Def.Process.Controller, Controller.pauseRecording))
+        self.wdgt.layout().addWidget(self._btn_pause, 1, 1)
+        ## Stop
+        self._btn_stop = QtWidgets.QPushButton('Stop')
+        self._btn_stop.clicked.connect(self.finalizeRecording)
+        self.wdgt.layout().addWidget(self._btn_stop, 1, 2)
+
+        ### Add buffers
+        self._cb_buffers = dict()
+        ## Camera buffers
+        self._grp_cameraBuffers = QtWidgets.QGroupBox('Camera buffers')
+        self._grp_cameraBuffers.setLayout(QtWidgets.QVBoxLayout())
+        self.wdgt.layout().addWidget(self._grp_cameraBuffers, 2, 0, 1, 3)
+        for bufferName in Config.Camera[Def.CameraCfg.buffers]:
+            bufferId = '{}/{}'.format(Def.Process.Camera, bufferName)
+            self._cb_buffers[bufferId] = QtWidgets.QCheckBox(bufferId)
+            self._cb_buffers[bufferId].clicked.connect(self.bufferStateChanged)
+            self._cb_buffers[bufferId].setTristate(False)
+            self._grp_cameraBuffers.layout().addWidget(self._cb_buffers[bufferId])
+        ## IO buffers
+        self._grp_ioBuffers = QtWidgets.QGroupBox('I/O buffers')
+        self._grp_ioBuffers.setLayout(QtWidgets.QVBoxLayout())
+        self.wdgt.layout().addWidget(self._grp_ioBuffers, 3, 0, 1, 3)
+        for bufferName in Config.IO[Def.IoCfg.buffers]:
+            bufferId = '{}/{}'.format(Def.Process.IO, bufferName)
+            self._cb_buffers[bufferId] = QtWidgets.QCheckBox(bufferId)
+            self._cb_buffers[bufferId].clicked.connect(self.bufferStateChanged)
+            self._cb_buffers[bufferId].setTristate(False)
+            self._grp_ioBuffers.layout().addWidget(self._cb_buffers[bufferId])
+
+        self.wdgt.layout().addItem(vSpacer, 4, 0)
+
+        ### Set timer for GUI update
+        self._tmr_updateGUI = QtCore.QTimer()
+        self._tmr_updateGUI.setInterval(100)
+        self._tmr_updateGUI.timeout.connect(self.updateGui)
+        self._tmr_updateGUI.start()
+
+    def finalizeRecording(self):
+        ### First: pause recording
+        IPC.rpc(Def.Process.Controller, Controller.pauseRecording)
+
+        reply = QtWidgets.QMessageBox.question(self, 'Finalize recording', 'Give me session data and stuff...',
+                                               QtWidgets.QMessageBox.Save | QtWidgets.QMessageBox.Discard,
+                                               QtWidgets.QMessageBox.Save)
+        if reply == QtWidgets.QMessageBox.Save:
+            print('Save metadata and stuff...')
+        else:
+            reply = QtWidgets.QMessageBox.question(self, 'Confirm discard', 'Are you sure you want to DISCARD all recorded data?',
+                                                   QtWidgets.QMessageBox.Yes | QtWidgets.QMessageBox.No,
+                                                   QtWidgets.QMessageBox.No)
+            if reply == QtWidgets.QMessageBox.Yes:
+                print('Fine... I`ll trash it all..')
+            else:
+                print('Puh... good choice')
+
+        ### Finally: stop recording
+        print('Stop recording...')
+        IPC.rpc(Def.Process.Controller, Controller.stopRecording)
+
+    def toggleEnable(self, newstate):
+        IPC.rpc(Def.Process.Controller, Controller.toggleEnableRecording, newstate)
+
+    def bufferStateChanged(self):
+        """Update shared buffer list for recordings based on UI selection.
+        Because buffer list is shared (managed) list, appending and removing has to be done in weird way (below)"""
+
+        for bufferId, cb in self._cb_buffers.items():
+            # Add bufferId if checkstate == True and bufferId not in buffer list
+            if Conversion.QtCheckstateToBool(cb.checkState()) and bufferId not in Config.Recording[Def.RecCfg.buffers]:
+                Config.Recording[Def.RecCfg.buffers] = Config.Recording[Def.RecCfg.buffers] + [bufferId]
+            # Remove bufferId if checkstate == False and bufferId is in buffer list
+            elif not(Conversion.QtCheckstateToBool(cb.checkState())) and bufferId in Config.Recording[Def.RecCfg.buffers]:
+                Config.Recording[Def.RecCfg.buffers] = [bi for bi in Config.Recording[Def.RecCfg.buffers] if bi != bufferId]
+
+    def updateGui(self):
+        """(Periodically) update UI based on shared configuration"""
+
+        enabled = Config.Recording[Def.RecCfg.enabled]
+        active = IPC.Control.Recording[Def.RecCtrl.active]
+        current_folder = IPC.Control.Recording[Def.RecCtrl.folder]
+
+
+        if active:
+            self.wdgt.setStyleSheet('QWidget#InGroupBox {background: rgba(179, 31, 18, 0.5);}')
+        else:
+            self.wdgt.setStyleSheet('QWidget#InGroupBox {background: rgba(0, 0, 0, 0.0);}')
+
+        ### Set enabled
+        self.setCheckable(not(active) and not(bool(current_folder)))
+        self.setChecked(enabled)
+
+        ### Set current folder
+        self._le_folder.setText(IPC.Control.Recording[Def.RecCtrl.folder])
+
+        ### Set buttons dis-/enabled
+        self._btn_start.setEnabled(not(active) and enabled)
+        self._btn_start.setText('Start' if IPC.inState(Def.State.IDLE, Def.Process.Controller) else 'Resume')
+        #self._btn_pause.setEnabled(active and enabled)
+        self._btn_pause.setEnabled(False)
+        self._btn_stop.setEnabled(bool(IPC.Control.Recording[Def.RecCtrl.folder]) and enabled)
+
+        ### Set buffer check states
+        for bufferId, cb in self._cb_buffers.items():
+            cb.setCheckState(Conversion.boolToQtCheckstate(bufferId in Config.Recording[Def.RecCfg.buffers]))
+
+        ### Enable/disable buffer groups during active recording
+        self._grp_cameraBuffers.setDisabled(active or not(enabled))
+        self._grp_ioBuffers.setDisabled(active or not(enabled))
+
+
+
 if __name__ == '__main__':
 
     if not('--skip_setup' in sys.argv):
