+++ buffers/IoBuffers.py
@@ -1,0 +1,4 @@
+from Buffer import AbstractBuffer
+
+class TestBuffer(AbstractBuffer):
+    pass
