+++ configs/default.ini
@@ -36,8 +36,8 @@ list_addons = EyePositionDetector
 bool_use = True
 str_device_type = Arduino
 str_device_model = ArduinoDue
-str_device_comport = COM7
-list_digital_pins = test01_pwmout:2:o:p:2,test02_digin:6:i::50
+str_device_comport = COM3
+list_digital_pins = test01_pwmout:2:p:2,test02_digin:6:i:50
 list_analog_pins = test01_anin:3:i:
 list_buffers = Analog,Digital
