--- gui/addons/__init__.py
@@ -1,2 +1,0 @@
-from .camera.EyePositionDetector import EyePositionDetector
-from .camera.TailDeflectionDetector import TailDeflectionDetector
